+++ tests/testthat/test-bm_glyph.R
@@ -1,8 +1,7 @@
 test_that("bm_glyph()", {
 
-    expect_error(bm_glyph("Zippity"),
-                 "Don't know how create 'bm_glyph' from this object"
-    )
+    expect_error(bm_glyph("Zippity"))
+    expect_error(bm_glyph(matrix("Zippity")))
 
     space_matrix <- matrix(0, nrow = 16, ncol = 16)
     space_glyph <- bm_glyph(space_matrix)
@@ -13,4 +12,7 @@ test_that("bm_glyph()", {
 
     space_glyph2 <- bm_glyph(space_glyph)
     expect_equal(space_glyph, space_glyph2)
+
+    space_glyph3 <- as_bm_glyph(space_matrix)
+    expect_equal(space_glyph, space_glyph3)
 })